Move library to updated package
