+++ .travis.yml
@@ -4,20 +4,14 @@ branches:
     - /^v4-\d+\.\d+\.\d+$/ # regex
 
 language: android
-sudo: false
+
 jdk:
     - oraclejdk8
 
-android:
-  components:
-    - tools
-    - platform-tools
-    - build-tools-27.0.3
-    - android-27
-    - doc-27
-
 before_install:
     - pip install --user codecov
+    - mkdir "$ANDROID_HOME/licenses" || true
+    - echo "d56f5187479451eabf01fb78af6dfcb131a6481e" > "$ANDROID_HOME/licenses/android-sdk-license"
 
 script:
   - ./gradlew clean testDebugUnitTest jacocoTestReport --info
@@ -30,11 +24,3 @@ cache:
   directories:
     - $HOME/.gradle
     - $HOME/.m2/repository
-
-deploy:
-  provider: script
-  script: ./gradlew bintrayUpload
-  skip_cleanup: true
-  on:
-    branch: master
-    tags: true
