+++ library/src/main/AndroidManifest.xml
@@ -6,8 +6,4 @@
   ~ LICENSE file in the root directory of this source tree. An additional grant
   ~ of patent rights can be found in the PATENTS file in the same directory.
   -->
-<manifest package="com.parse.facebook">
-
-    <application />
-
-</manifest>
+<manifest package="com.parse.facebook"/>
