+++ library/src/main/java/com/parse/facebook/ParseFacebookUtils.java
@@ -6,7 +6,7 @@
  * LICENSE file in the root directory of this source tree. An additional grant
  * of patent rights can be found in the PATENTS file in the same directory.
  */
-package com.parse;
+package com.parse.facebook;
 
 import android.app.Activity;
 import android.content.Context;
@@ -14,6 +14,11 @@ import android.content.Intent;
 import android.support.v4.app.Fragment;
 
 import com.facebook.AccessToken;
+import com.parse.AuthenticationCallback;
+import com.parse.LogInCallback;
+import com.parse.ParseException;
+import com.parse.ParseUser;
+import com.parse.SaveCallback;
 
 import java.util.Collection;
 import java.util.Collections;
