+++ library/src/test/java/com/parse/facebook/FacebookControllerTest.java
@@ -6,7 +6,7 @@
  * LICENSE file in the root directory of this source tree. An additional grant
  * of patent rights can be found in the PATENTS file in the same directory.
  */
-package com.parse;
+package com.parse.facebook;
 
 import android.app.Activity;
 import android.content.Context;
